v0.0.3

+++ src/consts.rs
@@ -13,6 +13,6 @@ pub const C3: &'static str = "B784B7";
 //
 
 pub const KEYWORDS: &[&str] = &[
-    "if", "else", "for", "while", "return", "int", "i16", "i32", "unt", "u16",
+    "if", "else", "for", "while", "return", "int", "i64", "i32", "unt", "u64",
     "u32", "flt", "f32", "f64", "bol", "str", "chr", "nil", "print", "im32", "im64", "im", "break"
 ];
